+++ packages/api-rx/src/subject.spec.js
@@ -2,14 +2,25 @@
 // This software may be modified and distributed under the terms
 // of the ISC license. See the LICENSE file for details.
 
-import createObservable from './subject';
+jest.mock('@polkadot/api-provider/ws', () => class {
+  isConnected = () => true;
+  on = () => true;
+  send = () => true;
+});
+
+const Api = require('./index').default;
 
 describe('subject', () => {
   const params = [123, false];
+  let api;
   let callback;
   let section;
   let subscription;
   let observable;
+
+  beforeEach(() => {
+    api = new Api();
+  });
 
   beforeEach(() => {
     const subMethod = jest.fn((name, ...params) => {
@@ -26,7 +37,7 @@ describe('subject', () => {
       subMethod
     };
 
-    observable = createObservable('subMethod', params, section);
+    observable = api.createSubject('subMethod', params, section);
   });
 
   afterEach(() => {
